refactor(suite): Add GlobalStyle for desktop app
