+++ packages/suite-desktop-ui/src/global.d.ts
@@ -4,12 +4,3 @@ interface Window {
     TrezorConnectIpcChannel?: TrezorConnectIpcChannel; // Electron API
     desktopFlags?: { exposeStore?: boolean };
 }
-
-declare global {
-    interface Window {
-        electronFind: {
-            onShow: (callback: () => void) => void;
-            offShow: (callback: () => void) => void;
-        };
-    }
-}
