+++ packages/suite/global.d.ts
@@ -40,4 +40,13 @@ declare module 'redux' {
     }
 }
 
+declare global {
+    interface Window {
+        electronFind: {
+            onShow: (callback: () => void) => void;
+            offShow: (callback: () => void) => void;
+        };
+    }
+}
+
 export {};
