+++ packages/suite/src/components/suite/FindBar/FindBar.tsx
