+++ packages/suite/src/components/suite/FindBar/highlight.ts
