+++ packages/suite/src/components/suite/FindBar/useFindBarShortcuts.ts
