+++ packages/suite/src/components/suite/labeling/MetadataLabeling/MetadataLabeling.tsx
@@ -20,9 +20,9 @@ import { MetadataAddPayload } from 'src/types/suite/metadata';
 import { LabelContentProps, LabelingVariant, MetadataProps, PrimitiveProps } from './definitions';
 import { withDropdown } from './withDropdown';
 import { withEditable } from './withEditable';
-import { NO_HIGHLIGHT_ATTRIBUTE } from '../../../../../../suite-desktop-ui/src/support/FindBar/consts';
 import { useLabelingCombined } from '../../../../hooks/suite/useLabelingCombined';
 import { AccountTypeBadge } from '../../AccountTypeBadge';
+import { NO_HIGHLIGHT_ATTRIBUTE } from '../../FindBar/consts';
 
 const LabelValue = styled.div`
     overflow: hidden;
