+++ packages/suite/src/support/suite/styles/GlobalStyle.tsx
@@ -3,10 +3,6 @@ import { DefaultTheme, createGlobalStyle } from 'styled-components';
 import { typography } from '@trezor/theme';
 
 import animations from './animations';
-import {
-    FIND_HIGHLIGHT_SELECTOR,
-    MARK_HIGHLIGHT_PULSE_SELECTOR,
-} from '../../../../../suite-desktop-ui/src/support/FindBar/consts';
 
 const GlobalStyle = createGlobalStyle<{ theme: DefaultTheme }>`
     #app {
@@ -59,30 +55,6 @@ const GlobalStyle = createGlobalStyle<{ theme: DefaultTheme }>`
     /* https://floating-ui.com/docs/misc#handling-large-content */
     .floating {
         max-width: calc(100vw - 10px);
-    }
-
-    ${FIND_HIGHLIGHT_SELECTOR} {
-        position: relative;
-        background: yellow;
-        color: black;
-        border-radius: 2px;
-        transition: transform 0.16s ease, background 0.16s ease;
-    }
-
-    ${FIND_HIGHLIGHT_SELECTOR}[data-active='true'] {
-        background: orange;
-        transform: scale(1.05);
-    }
-
-    ${FIND_HIGHLIGHT_SELECTOR} ${MARK_HIGHLIGHT_PULSE_SELECTOR} {
-        position: absolute;
-        inset: -3px;
-        border-radius: inherit;
-        background: orange;
-        opacity: 0.7;
-        transform: scale(1);
-        pointer-events: none;
-        z-index: -1;
     }
 `;
